Add custom Jackson serializer to REST API for primitive configurations.
